refactor: change skill proficiency type from double to int and adjust layout spacing in SkillCard for improved design consistency

+++ lib/models/skill.dart
@@ -4,7 +4,7 @@ class Skill {
   final String name;
   final String desc;
   final String tag;
-  final double percent;
+  final int percent;
   final Color color;
   final String image;
